+++ hosts/meowth/default.nix
@@ -2,10 +2,10 @@
 {
   imports = with self.nixosModules; [
     agenix
-    grafana-agent
+    alloy
     microvm-guest
-    prowlarr
     nix-topology
+    prowlarr
     time
     timesyncd
   ];